computed tools on index
